Module-14-Hands-on-Implementation/03-gemini-demo/requirements.txt
google-generativeai
streamlit
python-dotenv
